+++ geostore/tests/test_models/test_other_models.py
@@ -1,4 +1,5 @@
 from django.contrib.gis.geos import Point
+from django.db.utils import IntegrityError
 from django.test import TestCase
 from django.utils.text import slugify
 
@@ -44,3 +45,14 @@ class FeatureTestCase(TestCase):
                                              'name': 'toto'
                                          })
         self.assertEqual(feature.geom, Point(0, 0, srid=app_settings.INTERNAL_GEOMETRY_SRID))
+
+    def test_constraint_feature_empty_geom(self):
+        with self.assertRaises(IntegrityError):
+            Feature.objects.create(layer=self.layer_schema,
+                                   geom='POINT EMPTY',)
+
+    def test_constraint_feature_valid_geom(self):
+        with self.assertRaises(IntegrityError):
+            Feature.objects.create(layer=self.layer_schema,
+                                   geom='POLYGON((0 0, 1 1, 1 2, 1 1, 0 0))',)
+