+++ geostore/validators.py
@@ -1,5 +1,6 @@
 import jsonschema
 from django.core.exceptions import ValidationError
+from django.utils.translation import gettext_lazy as _
 from jsonschema.validators import validator_for
 
 
@@ -42,3 +43,11 @@ def validate_geom_type(layer_geom_type_id, feature_geom_type_id):
         if layer_geom_type_id != feature_geom_type_id:
             raise ValidationError(message='Geometry type is not the same on the layer')
     return feature_geom_type_id
+
+
+def validate_geom(feature_geom):
+    if feature_geom.empty:
+        raise ValidationError(_('Geometry is empty'))
+    if not feature_geom.valid:
+        raise ValidationError(_('Geometry is not valid'))
+    return feature_geom